paper_retriever/requirements.txt
unstructured[all-docs]==0.15.10
requests==2.32.3
python-mecab-ko==1.3.7
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2